fix: Home Screen work
Home screen working fine
