--- src/components/PracticeScreen.tsx
@@ -1,91 +1,0 @@
-// import {StyleSheet, Text, View} from 'react-native';
-// import React, {useEffect} from 'react';
-// import SplashScreen from 'react-native-splash-screen';
-// import * as Progress from 'react-native-progress';
-// import {SafeAreaView} from 'react-native-safe-area-context';
-
-// const PracticeScreen = () => {
-//   useEffect(() => {
-//     SplashScreen.hide();
-//   }, []);
-//   return (
-//     <SafeAreaView>
-//       {/* <Text>PracticeScreen</Text> */}
-//       {/* <Progress.Bar progress={0} width={200} /> */}
-
-//       {/* <Progress.Pie  progress={1} indeterminate={true}   size={200} /> */}
-//       <View style={styles.progresstwo}>
-//         <View style={styles.progresstwo}>
-//           <Progress.Circle
-//             size={300}
-//             indeterminate={true}
-//             endAngle={0.5}
-//             allowFontScaling={true}
-//             strokeCap="round"
-//             showsText={false}
-//             thickness={3}
-//             progress={1}
-//             borderWidth={1}
-//             direction="clockwise"
-//             animated={false}
-//             // backgroundColor="red"
-//             color="green"
-//             unfilledColor="green"
-//             indeterminateAnimationDuration={10000}
-//           />
-
-//           <Progress.Circle
-//             size={300}
-//             indeterminate={true}
-//             endAngle={0.5}
-//             allowFontScaling={true}
-//             strokeCap="round"
-//             showsText={false}
-//             thickness={3}
-//             progress={1}
-//             borderWidth={1}
-//             direction="clockwise"
-//             animated={false}
-//             // backgroundColor="red"
-//             color="green"
-//             unfilledColor="green"
-//             indeterminateAnimationDuration={10000}
-//           />
-//           {/* <View style={styles.progresstwo}> */}
-//           {/* <Progress.Circle
-//         size={330}
-//         indeterminate={true}
-//         endAngle={.5}
-//         allowFontScaling={true}
-//         strokeCap="round"
-//         showsText={false}
-//         thickness={3}
-//         progress={1}
-//         borderWidth={1}
-//         direction="clockwise"
-//         animated={false}
-//         backgroundColor="red"
-//         color="green"
-//         indeterminateAnimationDuration={10000}
-//       /> */}
-//           {/* </View> */}
-//         </View>
-//       </View>
-
-//       {/* <Progress.CircleSnail color={['red', 'green', 'blue']} /> */}
-//     </SafeAreaView>
-//   );
-// };
-
-// export default PracticeScreen;
-
-// const styles = StyleSheet.create({
-//   progresstwo: {
-//     height: 300,
-//     width: 300,
-//     backgroundColor: 'white',
-//     borderRadius: 150,
-//     marginTop: '1%',
-//     // marginTop:
-//   },
-// });
