--- src/redux/demo/update.ts
@@ -1,14 +1,0 @@
-import {createSlice, PayloadAction} from '@reduxjs/toolkit';
-
-const counterSlice = createSlice({
-  name: 'update',
-  initialState: 0,
-  reducers: {
-    increment: (state, action: PayloadAction<number>) => state + action.payload,
-    decrement: (state, action: PayloadAction<number>) => state - action.payload,
-    setCustomValue: (state, action: PayloadAction<number>) => action.payload,
-  },
-});
-
-export const {increment, decrement, setCustomValue} = counterSlice.actions;
-export default counterSlice.reducer;
