--- src/redux/index.ts
@@ -1,15 +1,0 @@
-import {configureStore} from '@reduxjs/toolkit';
-import {useDispatch, useSelector, TypedUseSelectorHook} from 'react-redux';
-import rootReducer from './rootReducer';
-
-export type RootState = ReturnType<typeof rootReducer>;
-
-const store = configureStore({
-  reducer: rootReducer,
-  // middleware: getDefaultMiddleware => getDefaultMiddleware(), // .prepend(middleware)
-});
-
-export type AppDispatch = typeof store.dispatch;
-export const useReduxDispatch = (): AppDispatch => useDispatch<AppDispatch>();
-export const useReduxSelector: TypedUseSelectorHook<RootState> = useSelector;
-export default store;
