--- src/redux/rootReducer.ts
@@ -1,9 +1,0 @@
-import {combineReducers} from '@reduxjs/toolkit';
-import counterReducer from './demo/update';
-
-const rootReducer = combineReducers({
-  // state
-  counter: counterReducer,
-});
-
-export default rootReducer;
